Move generate_results_table_from_vector to vector_utils; remove sdr_ndr_utils; rename test_sdr_ndr_utils to test_vector_utils.
